+++ src/App.css
@@ -1,89 +1,42 @@
-
-
-
-
-@import url('https://fonts.googleapis.com/css?family=Raleway:400,700');
-
-
-* {
-	box-sizing: border-box;
-	margin: 0;
-	padding: 0;	
-	font-family: Raleway, sans-serif;
-}
-h1 {
-  color: rgb(242, 62, 92);
-}
-body{
-  padding: 0;
-  position: relative;
-  background: #fff;
+#root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
 }
 
-.container {
-  position: absolute;
-  left: 0;
-  top: 0;
-  background: linear-gradient(90deg, #C7C5F4, #776BCC);		
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	height: 35vh;
-  width: 100vw;
-  margin-top: 0;
+.logo {
+  height: 6em;
+  padding: 1.5em;
+  will-change: filter;
+  transition: filter 300ms;
 }
-.fullContainer{
-  position: absolute;
-  left: 0;
-  top: 0;
-  background: linear-gradient(90deg, #C7C5F4, #776BCC);		
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	height: 100vh;
-  width: 100vw;
-  margin-top: 0;
+.logo:hover {
+  filter: drop-shadow(0 0 2em #646cffaa);
 }
-.jokeTitle{
-  height: 65vh;
-  width: 100vw;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: #776BCC;
-  font-size: 5rem;
+.logo.react:hover {
+  filter: drop-shadow(0 0 2em #61dafbaa);
 }
-.icon{
-  position: absolute;
-  color: #000;
-  font-size: 2rem;
-  cursor: pointer;
-}
-.icon.up{
-  animation: 0.5s infinite alternate sliding;
-}
-.icon.down{
-  animation: 0.5s infinite alternate slidingBack;
-}
-.icon:hover{
-  opacity: 0.7;
-}
-@keyframes sliding {
+
+@keyframes logo-spin {
   from {
-    top: 32vh;
+    transform: rotate(0deg);
   }
-
   to {
-   top: 33vh;
-  }
-}
-@keyframes slidingBack {
-  from {
-    top: 90vh;
-  }
-
-  to {
-   top: 88vh;
+    transform: rotate(360deg);
   }
 }
 
+@media (prefers-reduced-motion: no-preference) {
+  a:nth-of-type(2) .logo {
+    animation: logo-spin infinite 20s linear;
+  }
+}
+
+.card {
+  padding: 2em;
+}
+
+.read-the-docs {
+  color: #888;
+}
